backend/requirements.txt
scrapy
flask
googlesearch-python
yfinance
pandas
requests
firecrawl
pydantic
python-dotenv
groq
langchain
neo4j
langchain-community 
langchain-core 
langchain-openai 
langchain-text-splitters 
tiktoken
langchain_groq
py2neo
matplotlib
networkx
python-dotenv